FREEI-1959 - optimize the queues_get() function
Previously, when fetching Destinations, we needed to fetch the edit_url
from the more detailed getModuleDestinationInfo() method. Unfortunatly,
in some situations, calling the getModuleDestinationInfo() method is
rather resource intensive because need to query from the Asterisk AGI
in order to provide the information requested. Luckily for us, the
edit_url information was the only information needed from the route
when fetching all Destinations, so, if the edit_url is provided in
the results, then there is no need to query getModuleDestinationInfo()
and an over 60% preformance gain is acheived.